==> 6a_var_id.R <==
var_desc <- c(
  "Total population [Mln pers]"
  ,    "Total GDP [Bln USD 2005]"
  ,    "Total fuel consumption [PJ]"
  ,    "Mandated bioenergy [PJ, %]"
  ,    "Real producer price [USD/t or USD/GJ] "
  ,    "Real producer price [USD/t or USD/GJ] "
  ,    "Real producer price index [USD/t or USD/GJ]   (aggregated as Laspeyres)"
  ,    "Real consumer price index [USD/t or USD/GJ or USD/1000 kcal] (aggregated as Laspeyres)"
  ,    "Area cultivated [1000 ha]"
  ,    "Area cultivated - rainfed [1000 ha]"
  ,    "Area cultivated - irrigated [1000 ha]"
  ,    "Area harvested [1000 ha]"
  ,    "Land cover [Mha]"
  ,    "Land rent [USD/ha]"
  ,    "Crop yield [t/ha]"
  ,    "Crop yield, with multicropping [t/ha]"
  ,    "Crop yield - rainfed [t/ha]"
  ,    "Crop yield - irrigated [t/ha]"
  ,    "Exogenous crop yield [t/ha]"
  ,    "Exogenous crop yield without CC [t/ha, %]"
  ,    "Land productivity (endogenous) [kg prot/ha]"
  ,    "Exogenous land productivity trend [kg prot/ha]"
  ,    "Feed productivity (endogenous) [kg prot/t dm feed]"
  ,    "Exogenous feed productivity trend [kg prot/t dm feed]"
  ,    "Food use [1000 t]"
  ,    "Feed use [1000 t]"
  ,    "Animal number [1000 TLU]"
  ,    "Animal feed intake [1000 t dm]"
  ,    "Biofuel use [1000 t]"
  ,    "Other use [1000 t]"
  ,    "Share used for biofuel production [%]"
  ,    "Implied biofuel blending rate (energy %)"
  ,    "Imports [1000 t]"
  ,    "Exports [1000 t]"
  ,    "Fertiliser N [1000 t]"
  ,    "Fertiliser P [1000 t]"
  ,    "Water for irrigation [km3]"
  ,    "p.c. calory availability [kcal/cap/d]"
  ,    "p.c. calory diet target [kcal/cap/d]"
  ,    "Emissions from agriculture [MtCO2eq]"
  ,    "CO2 emissions [MtCO2eq]"
  ,    "CH4 emissions [MtCO2eq]"
  ,    "N2O emissions [MtCO2eq]"
  ,    "Non CO2 emissions [MtCO2eq]"
  ,    "Production [1000 t, PJ]"
  ,    "Domestic use [1000 t]"
  ,    "Net trade [1000 t]"
  ,    "Net trade share in market volume [%]"
  ,     "Carbon tax [USD/tCO2]"
)
var_id <- c( "POPT" 
             ,"GDPT"
             ,"FUEL"
             ,"QBFL"
             ,"XPRP"
             ,"XPRX"
             ,"XPRI"
             ,"XCPI"
             ,"AREA"
             ,"ARRF"
             ,"ARIR"
             ,"HARV"
             ,"LAND"
             ,"LRNT"
             ,"YILD"
             ,"YILM"
             ,"YIRF"
             ,"YIIR"
             ,"YEXO"
             ,"YECC"
             ,"LYLD"
             ,"LYXO"
             ,"FEEF"
             ,"FEXO"
             ,"FOOD"
             ,"FEED"
             ,"ANIM"
             ,"ANFD"
             ,"BIOU"
             ,"OTHU"
             ,"BFSH"
             ,"BFBL"
             ,"IMPO"
             ,"EXPO"
             ,"FRTN"
             ,"FRTP"
             ,"WATR"
             ,"CALO"
             ,"CALT"
             ,"EMIS"
             ,"ECO2"
             ,"ECH4"
             ,"EN2O"
             ,"ENCO"
             ,"PROD"
             ,"CONS"
             ,"NETT"
             ,"NTMS"
             ,"CTAX")
var_id <- bind_cols(var_id,var_desc)
names(var_id) <- c("VAR_ID","VAR_DESC")

# globiom_path <- "C:/Users/Scott/Documents/_PhD/GLOBIOM_Prerelease_Model-master"# globiom_path <- "C:/Users/Scott/Documents/_PhD/GLOBIOM_Prerelease_Model-master"
# # globiom_run <- file.path(globiom_path,"Model/0_executebatch.gms")
# # gams(globiom_run)
 globiom_path <- here("Inputs/")
globiom_file <- file.path(globiom_path, "Model/output/Baseline_MyRun.gdx")
